pyproject.toml
[tool.poetry]
name = "zentric-project"
version = "0.1.0"
description = ""
authors = ["Yasser Azan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
drf-yasg = "^1.21.7"
boto3 = "^1.34.151"
graphene-django = "^3.2.2"
faker = "^26.0.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
